Move graphite parser to separate file
